Reorganize code base with folders. Make utils directory a package by adding __init__.py
