+++ misc/ideas.md
